+++ featuretools/tests/primitive_tests/test_transform_features.py
@@ -741,153 +741,6 @@ def test_text_primitives(es):
     for i, v in enumerate(char_values):
         assert v == char_counts[i]
 
-# M TODO
-    # value2 + 2,
-    # value2 - 2,
-    # value2 * 2,
-    # value2 / 2,
-    # value2 > 2,
-    # value2 < 2,
-    # value2 == 2,
-    # value2 != 2,
-    # value2 >= 2,
-    # value2 <= 2,
-    #                 (value2, 2)
-    #     our_reverse_overrides = [
-    #     2 + value2,
-    #     2 - value2,
-    #     2 * value2,
-    #     2 / value2]
-    # i = 0
-    # for feat in feats:
-    #     if feat != Mod:
-    #         f = feat(2, value2)
-    #         o = our_reverse_overrides[i]
-    #         assert o.hash() == f.hash()
-    #         i += 1
-
-    # python_reverse_overrides = [
-    #     2 < value2,
-    #     2 > value2,
-    #     2 == value2,
-    #     2 != value2,
-    #     2 <= value2,
-    #     2 >= value2]
-    # i = 0
-    # for compare_op in compare_ops:
-    #     f = compare_op(value2, 2)
-    #     o = python_reverse_overrides[i]
-    #     assert o.hash() == f.hash()
-    #     i += 1
-
-
-def test_overrides(es):
-    value = ft.Feature(es['log']['value'])
-    value2 = ft.Feature(es['log']['value_2'])
-
-    feats = [AddNumeric, SubtractNumeric, MultiplyNumeric, DivideNumeric, GreaterThan,
-             LessThan, Equal, NotEqual, GreaterThanEqualTo, LessThanEqualTo]
-    assert ft.Feature(value, primitive=Negate()).hash() == (-value).hash()
-
-    compares = [(value, value), (value, value2)]
-    overrides = [
-        value + value,
-        value - value,
-        value * value,
-        value / value,
-        value > value,
-        value < value,
-        value == value,
-        value != value,
-        value >= value,
-        value <= value,
-
-        value + value2,
-        value - value2,
-        value * value2,
-        value / value2,
-        value > value2,
-        value < value2,
-        value == value2,
-        value != value2,
-        value >= value2,
-        value <= value2,
-    ]
-
-    for left, right in compares:
-        for feat in feats:
-            f = ft.Feature([left, right], primitive=feat)
-            o = overrides.pop(0)
-            assert o.hash() == f.hash()
-
-
-def test_override_boolean(es):
-    count = ft.Feature(es['log']['id'], parent_entity=es['sessions'], primitive=Count)
-    count_lo = ft.Feature(count,primitive=GreaterThanScalar(1))
-    count_hi = ft.Feature(count, primitive=LessThanScalar(10))
-
-    to_test = [[True, True, True],
-               [True, True, False],
-               [False, False, True]]
-
-    features = []
-    features.append(count_lo.OR(count_hi))
-    features.append(count_lo.AND(count_hi))
-    features.append(~(count_lo.AND(count_hi)))
-
-    pandas_backend = PandasBackend(es, features)
-    df = pandas_backend.calculate_all_features(instance_ids=[0, 1, 2],
-                                               time_last=None)
-    for i, test in enumerate(to_test):
-        v = df[features[i].get_name()].values.tolist()
-        assert v == test
-
-
-def test_override_cmp_from_variable(es):
-    count_lo = ft.Feature(es['log']['value']) > 1
-
-    to_test = [False, True, True]
-
-    features = [count_lo]
-
-    pandas_backend = PandasBackend(es, features)
-    df = pandas_backend.calculate_all_features(instance_ids=[0, 1, 2],
-                                               time_last=None)
-    v = df[count_lo.get_name()].values.tolist()
-    for i, test in enumerate(to_test):
-        assert v[i] == test
-
-
-def test_override_cmp(es):
-    count = ft.Feature(es['log']['id'], parent_entity=es['sessions'], primitive=Count)
-    _sum = ft.Feature(es['log']['value'], parent_entity=es['sessions'], primitive=Sum)
-    gt_lo = count > 1
-    gt_other = count > _sum
-    ge_lo = count >= 1
-    ge_other = count >= _sum
-    lt_hi = count < 10
-    lt_other = count < _sum
-    le_hi = count <= 10
-    le_other = count <= _sum
-    ne_lo = count != 1
-    ne_other = count != _sum
-
-    to_test = [[True, True, False],
-               [False, False, True],
-               [True, True, True],
-               [False, False, True],
-               [True, True, True],
-               [True, True, False],
-               [True, True, True],
-               [True, True, False]]
-    features = [gt_lo, gt_other, ge_lo, ge_other, lt_hi,
-                lt_other, le_hi, le_other, ne_lo, ne_other]
-    pandas_backend = PandasBackend(es, features)
-    df = pandas_backend.calculate_all_features(instance_ids=[0, 1, 2],
-                                               time_last=None)
-    for i, test in enumerate(to_test):
-        v = df[features[i].get_name()].values.tolist()
-        assert v == test
 
 
 def test_isin_feat(es):
